+++ docs/changelog.rst
@@ -13,6 +13,8 @@ Changelog
   with a deprecation warning.  This will be removed in 2.0.
 - Add ``normalize_parameter_names`` keyword to the
   :func:`headers._parse_parameter_list` internal function.
+- Add support for parsing :rfc:`7239` ``Forwarded`` headers with
+  :func:`headers.parse_forwarded`.
 
 `1.4.3`_ (30-Oct-2017)
 ----------------------